fix(plugins/tech-radar): list rerenders on hover
