+++ plugins/tech-radar/src/components/RadarLegend/types.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2020 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Entry, Quadrant, Ring } from "../../utils/types";
+
+export type Segments = {
+  [k: number]: { [k: number]: Entry[] };
+};
+
+export type RadarLegendProps = {
+  quadrants: Quadrant[];
+  rings: Ring[];
+  entries: Entry[];
+  onEntryMouseEnter?: (entry: Entry) => void;
+  onEntryMouseLeave?: (entry: Entry) => void;
+};
